✨ refactor logging and utilities organization
- Modularized isQuiet to core for better code reuse.
