+++ packages/core/src/quiet.ts
@@ -1,0 +1,11 @@
+// Boolean indicating if debug messages should be suppressed
+// Controls whether debug messages are outputted
+export let isQuiet = false
+
+/**
+ * Sets the quiet mode for suppressing debug messages.
+ * @param v - Boolean to enable or disable quiet mode
+ */
+export function setQuiet(v: boolean) {
+    isQuiet = !!v
+}
